fix(tools): move repoRoot out of the integration build tag   fix(githooks): replace declare -A so pre-commit runs on macOS bash 3.2   fix(lint): lint the autoshift chart with its values instead of bare Signed-off-by: Ben Carr <9310348+bcarr-rh@users.noreply.github.com>
